apps/api/pyproject.toml
[project]
name = "lab-notebook-api"
version = "0.1.0"
description = "Backend API for Lab Notebook"
requires-python = ">=3.12"
dependencies = [
  "fastapi[standard]",
  "sqlalchemy>=2.0.0",
  "alembic>=1.13.0",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.0.0",
  "python-multipart>=0.0.9",
  "pyjwt>=2.8.0",
  "passlib[bcrypt]>=1.7.4"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]